Trabajmos en agregrar la id de pagoque pasa MP por IPN al DBA

+++ Class_MP/connect_db.py
@@ -1,5 +1,7 @@
 import pypyodbc
 import json
+import time
+
 
 class ConexionSybase:
     def __init__(self, **kwargs):
@@ -45,10 +47,11 @@ class ConexionSybase:
             self.conectar()
             query = f"""
                 CREATE TABLE {nombre_tabla} (
-                    ID INT PRIMARY KEY IDENTITY,
-                    NRO_FACTURA VARCHAR(255),
+                    ID INT IDENTITY,
+                    FECHA_CREACION DATETIME,
+                    NRO_FACTURA VARCHAR(255) PRIMARY KEY,
+                    PUNTO_VENTA VARCHAR(255),
                     ID_PAGO VARCHAR(255),
-                    ID_COMPRA BIGINT
                 )
             """
             self.cursor.execute(query)
@@ -58,6 +61,8 @@ class ConexionSybase:
             print(f'Error al crear la tabla: {e}')
         finally:
             self.desconectar()
+
+
 
     def mostrar_tablas(self, nombre_bda):
         try:
@@ -83,15 +88,35 @@ class ConexionSybase:
                 print(f"Tabla '{nombre_tabla}' actualizada exitosamente.")
         except pypyodbc.Error as err:
             print(f"Error al actualizar la tabla: {err}")
+            
+    def verificar_existencia_tabla(self, nombre_tabla):
+        try:
+            with self.conexion.cursor() as cursor:
+                consulta = f"SELECT COUNT(*) FROM SYSTABLE WHERE TABLE_NAME = '{nombre_tabla}'"
+                cursor.execute(consulta)
+                cantidad = cursor.fetchone()[0]
+                return cantidad > 0
+        except pypyodbc.Error as err:
+            print(f"Error al verificar la existencia de la tabla: {err}")
+            return False
+
 
     def eliminar_tabla(self, nombre_tabla):
         try:
-            with self.conexion.cursor() as cursor:
-                consulta = f"DROP TABLE {nombre_tabla}"
-                cursor.execute(consulta)
-                print(f"Tabla '{nombre_tabla}' eliminada exitosamente.")
+            if self.conectar():  # Llama a conectar antes de continuar
+                if self.verificar_existencia_tabla(nombre_tabla):
+                    with self.conexion.cursor() as cursor:
+                        consulta = f"DROP TABLE {nombre_tabla}"
+                        cursor.execute(consulta)
+                        print(f"Tabla '{nombre_tabla}' eliminada exitosamente.")
+                else:
+                    print(f"La tabla '{nombre_tabla}' no existe en la base de datos.")
+            else:
+                print("Error al conectar a la base de datos.")
         except pypyodbc.Error as err:
             print(f"Error al eliminar la tabla: {err}")
+
+
             
     def limpiar_tabla(self, tabla):
         try:
@@ -116,9 +141,10 @@ class ConexionSybase:
         except pypyodbc.Error as err:
             print(f"Error al seleccionar la tabla '{nombre_tabla}': {err}")            
 
-    #MANEJO DE DATOS DE UNA TABLA
+    # MANEJO DE DATOS DE UNA TABLA
     def insertar_datos(self, tabla, datos):
         try:
+            self.conectar()
             with self.conexion.cursor() as cursor:
                 # Construir la consulta SQL de inserción
                 columnas = ", ".join([f'"{col}"' for col in datos.keys()])
@@ -129,55 +155,80 @@ class ConexionSybase:
                 print(f"Datos insertados en la tabla '{tabla}' exitosamente.")
         except pypyodbc.Error as err:
             print(f"Error al insertar datos: {err}")
-            
-    def insertar_datos_MERCHANT(self, NRO_FACTURA):
-        try:
-            self.conectar()
-            datos = {
-                'NRO_FACTURA': NRO_FACTURA,
-            }
-            self.insertar_datos('MERCHANTORDEN', datos)
-        except Exception as e:
-            print(f'Error al insertar datos en MERCHANTORDEN: {e}')
         finally:
             self.desconectar()
+
+    def insertar_datos_MERCHANT(self, NRO_FACTURA, PUNTO_VENTA):
+        datos = {
+            'NRO_FACTURA': NRO_FACTURA,
+            'PUNTO_VENTA': PUNTO_VENTA
+        }
+        self.insertar_datos('MERCHANTORDEN', datos)
+
             
-    def actualizar_id_pago(self, numero_factura, ID_PAGO):
+    def actualizar_id_pago(self, factura, POS, ID_PAGO):
         try:
             self.conectar()
 
-            with self.conexion.cursor() as cursor:
-                # Ejecutar la consulta para actualizar el ID_PAGO
-                query = f"UPDATE MERCHANTORDEN SET ID_PAGO = '{ID_PAGO[0]}' WHERE NRO_FACTURA = '{numero_factura}'"
-                cursor.execute(query)
-                self.conexion.commit()
+            # Verificar si existe una fila con la combinación de NRO_FACTURA y PUNTO_VENTA
+            if self.existe_fila(factura, POS):
+                with self.conexion.cursor() as cursor:
+                    # Ejecutar la consulta para actualizar el ID_PAGO
+                    query = f"UPDATE MERCHANTORDEN SET ID_PAGO = '{ID_PAGO[0]}' WHERE NRO_FACTURA = '{factura}' AND PUNTO_VENTA = '{POS}'"
+                    cursor.execute(query)
+                    self.conexion.commit()
 
-                print(f"ID_PAGO actualizado exitosamente para la factura {numero_factura}.")
+                    print(f"ID_PAGO actualizado exitosamente para la factura {POS}.")
+            else:
+                print(f"No se encontró una fila con NRO_FACTURA='{factura}' y PUNTO_VENTA='{POS}' en la base de datos.")
         except pypyodbc.Error as err:
             print(f"Error al actualizar ID_PAGO: {err}")
         finally:
             self.desconectar()
-    
-    # MANEJO DE DATOS DE UNA TABLA
-    def obtener_id_compra(self, numero_factura):
+
+
+    def existe_fila(self, numero_factura, punto_venta):
         try:
             with self.conexion.cursor() as cursor:
-                # Ejecutar la consulta para obtener el ID de compra
-                query = f"SELECT ID_COMPRA FROM MERCHANTORDEN WHERE NRO_FACTURA = '{numero_factura}'"
+                # Ejecutar la consulta para verificar la existencia de la fila
+                query = f"SELECT 1 FROM MERCHANTORDEN WHERE NRO_FACTURA = '{numero_factura}' AND PUNTO_VENTA = '{punto_venta}'"
                 cursor.execute(query)
 
                 # Obtener el resultado de la consulta
                 resultado = cursor.fetchone()
 
-                if resultado:
-                    # Si se encontró un resultado, devolver el ID de compra
-                    return resultado[0]
+                # Si se encontró un resultado, la fila existe
+                return resultado is not None
+        except pypyodbc.Error as err:
+            print(f"Error al verificar la existencia de la fila: {err}")
+            return False
+
+    
+    # MANEJO DE DATOS DE UNA TABLA
+    def obtener_id_compra(self, numero_factura, punto_venta):
+        try:
+            with self.conexion.cursor() as cursor:
+                # Ejecutar la consulta para obtener el ID de compra y ID_PAGO
+                query = f"SELECT ID_PAGO FROM MERCHANTORDEN WHERE NRO_FACTURA = '{numero_factura}' AND PUNTO_VENTA = '{punto_venta}'"
+                cursor.execute(query)
+
+                # Obtener el resultado de la consulta
+                resultado = cursor.fetchone()
+
+                if resultado is not None:
+                    # Si se encontró un resultado, devolver el ID de compra y ID_PAGO
+                    id_pago = resultado
+                    print(f"ID_PAGO para factura {numero_factura} y POS {punto_venta}: {id_pago}")
+                    return id_pago
                 else:
-                    # Si no se encontró ninguna coincidencia, devolver None o algún valor indicativo
-                    return 
+                    # Si no se encontró ninguna coincidencia, devolver None
+                    print(f"No se encontró una fila con NRO_FACTURA='{numero_factura}' y PUNTO_VENTA='{punto_venta}'.")
+                    return None
         except pypyodbc.Error as err:
             print(f"Error al obtener el ID de compra: {err}")
             return None
+
+
     #DESCONEXION DE LA BASE DE DATOS
     def desconectar(self):
         if self.conexion and self.conexion.connected:
